requirements.txt
mcp>=1.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
ytmusicapi>=1.7.0
pytube>=15.0.0
requests>=2.31.0
flask>=3.0.0
